pyproject.toml
[project]
name = "calm"
version = "0.1.0"
description = "A CLI tool to interact with Google Calendar."
requires-python = ">=3.9"
dependencies = [
  "typer>=0.12.3",
  "google-api-python-client>=2.136.0",
  "google-auth-oauthlib>=1.2.1",
  "google-auth-httplib2>=0.2.0",
  "tabulate>=0.9.0",
]

[project.scripts]
calm = "calm.cli:app"